Additional VS 2022 C++17 compatibility fixes
Fixed two critical compilation issues for Visual Studio 2022:

1. Thread library std::unique_ptr initialization (thread.cpp:105)
   - Changed from explicit unique_ptr constructor to std::make_unique
   - MSVC 2022 has stricter requirements for unique_ptr initialization
   - std::make_unique is the modern C++14 idiom and works correctly

2. SFINAE template resolution (class.h:251-256)
   - Made enable_if template parameters distinct using different types
   - Changed false-case from 'int' to 'long' to help MSVC disambiguate
   - Fixes "no matching overloaded function found" for get_unserialize_functor
   - Resolves Stream class registration (MTS_IMPLEMENT_CLASS macro)

These changes maintain compatibility with GCC/Clang while enabling
successful compilation with MSVC 2022's stricter template deduction.

+++ mitsuba/include/mitsuba/core/class.h
@@ -248,11 +248,11 @@ constexpr bool is_constructible_v = is_constructiblee<T, Arg>::value;
 
 template <typename T, std::enable_if_t<is_constructible_v<T, const Properties&>, int> = 0>
 Class::ConstructFunctor get_construct_functor() { return [](const Properties& p) -> Object* { return new T(p); }; }
-template <typename T, std::enable_if_t<!is_constructible_v<T, const Properties&>, int> = 0>
+template <typename T, std::enable_if_t<!is_constructible_v<T, const Properties&>, long> = 0>
 Class::ConstructFunctor get_construct_functor() { return {}; }
 template <typename T, std::enable_if_t<is_constructible_v<T, Stream*>, int> = 0>
 Class::UnserializeFunctor get_unserialize_functor() { return [](Stream* s) -> Object* { return new T(s); }; }
-template <typename T, std::enable_if_t<!is_constructible_v<T, Stream*>, int> = 0>
+template <typename T, std::enable_if_t<!is_constructible_v<T, Stream*>, long> = 0>
 Class::UnserializeFunctor get_unserialize_functor() { return {}; }
 NAMESPACE_END(detail)
 NAMESPACE_END(mitsuba)
